refactor(cockpit): extract trend module
